requirements.txt
alchemy==20.5
alembic==1.13.1
blinker==1.7.0
Brotli==1.1.0
cachelib==0.13.0
cairocffi==1.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
click==8.1.7
cssselect2==0.7.0
docutils==0.21.1
filelock==3.0.12
Flask==3.0.3
flask-form==1.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
flask_weasyprint==1.1.0
fonttools==4.51.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
gunicorn==20.1.0
haversine==2.8.1
html5lib==1.1
idna==2.8
itsdangerous==2.2.0
Jinja2==3.1.3
lockfile==0.12.2
Mako==1.3.3
MarkupSafe==2.1.5
oauthlib==2.1.0
pdfkit==1.0.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.10.0
pyphen==0.15.0
pypng==0.20220715.0
python-daemon==2.2.4
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
requests==2.22.0
requests-oauthlib==1.1.0
setuptools==69.5.1
six==1.16.0
sql==2022.4.0
SQLAlchemy==2.0.29
tinycss2==1.3.0
typing_extensions==4.11.0
urllib3==1.25.11
weasyprint==62.1
webencodings==0.5.1
Werkzeug==3.0.2
zopfli==0.2.3
